Move CharacterKey to ID list
